+++ src/App.jsx
@@ -1,4 +1,4 @@
-import './App.css';
+import './styles/App.css';
 import Header from './components/Header';
 import Members from './components/Members';
