+++ src/index.jsx
@@ -5,7 +5,7 @@ import Particles from 'particlesjs';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import './index.css';
+import './styles/index.css';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
