+++ src/styles/App.css
@@ -1,0 +1,421 @@
+html,
+body {
+  margin: 0;
+  padding: 0;
+}
+
+.background {
+  position: absolute;
+  /* display: block; */
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: -1;
+}
+
+body {
+  color: white;
+  /* background-image: url("./original.jpg"); */
+  /* background-size: 100%;
+  background-repeat: repeat; */
+  align-items: center;
+  margin: 0% 5% 0% 5%;
+}
+
+.container-fluid {
+  position: relative;
+  padding: 1% 1% 1% 1%;
+}
+
+.header-element {
+  margin: 5%;
+  padding: auto;
+}
+
+.content {
+  text-align: center;
+  background-color: rgba(0, 0, 0, 1);
+  margin: 0% 10% 0% 10%;
+  padding: 7%;
+  border-radius: 2rem;
+}
+
+.header-element .content {
+  background-color: rgba(0, 0, 0, 0.7);
+}
+
+.logo-div img {
+  display: block;
+  display: relative;
+  margin-left: auto;
+  margin-right: auto;
+  margin-bottom: 0%;
+  border-radius: 100%;
+  width: 25vh;
+}
+
+.header-element h1 {
+  text-align: center;
+  margin-top: 5%;
+  margin-bottom: 0%;
+  font-size: 5.5rem;
+  color: orange;
+}
+
+h1 {
+  text-align: center;
+  margin-top: 5%;
+  margin-bottom: 0%;
+  font-size: 2.5rem;
+  color: orange;
+}
+
+.header-element .content h1 {
+  font-size: 4.5vh;
+}
+
+.content .description {
+  margin-top: 5%;
+  font-size: 1rem;
+}
+.content .socials img {
+  border-radius: 100%;
+  background-color: grey;
+  padding: 5px 5px 5px 5px;
+  margin: 0% 5% 0% 5%;
+  width: 75px;
+}
+
+.members-content {
+  background-color: black;
+  margin: 0% 10% 0% 10%;
+  padding: 7%;
+  border-radius: 15px;
+}
+
+@media (min-width: 400px) {
+  .logo-div img {
+    width: 30vh;
+  }
+
+  .content .description {
+    font-size: 2rem;
+  }
+  .content .socials img {
+    width: 75px;
+  }
+}
+
+@media (max-width: 1000px) {
+  .logo-div img {
+    width: 20vh;
+  }
+
+  .content .description {
+    font-size: 1rem;
+  }
+
+  .content .socials img {
+    width: 5vh;
+  }
+}
+
+.button {
+  margin: 0%;
+  background-color: rgba(0, 0, 0, 0);
+  width: 300px;
+  z-index: 0;
+}
+
+/* Members Select Module*/
+
+* {
+  box-sizing: border-box;
+  font-family: 'Roboto', sans-serif;
+  text-shadow: -1px -1px 0 #000, 1px -1px 0 #000, -1px 1px 0 #000, 1px 1px 0 #000;
+}
+.honeycomb {
+  display: -webkit-box;
+  display: flex;
+  flex-wrap: wrap;
+  list-style: none;
+  -webkit-box-pack: center;
+  justify-content: center;
+  -webkit-box-align: center;
+  align-items: center;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0;
+}
+
+.honeycomb-cell {
+  -webkit-box-flex: 0;
+  flex: 0 1 250px;
+  max-width: 250px;
+  height: 137.5px;
+  margin: 65.4px 12.5px 25px;
+  position: relative;
+  padding: 0.5em;
+  text-align: center;
+  /* z-index: 0; */
+  box-shadow: 0px 0px 15px 0 rgba(0, 0, 0, 0.1);
+}
+.honeycomb-cell_img {
+  object-fit: cover;
+  object-position: center;
+  filter: grayscale(50%);
+}
+.honeycomb-cell_title {
+  height: 100%;
+  display: -webkit-box;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+  flex-direction: column;
+  -webkit-box-pack: center;
+  justify-content: center;
+  -webkit-hyphens: auto;
+  hyphens: auto;
+  word-break: break-word;
+  text-transform: uppercase;
+  color: #fff;
+  font-weight: 700;
+  font-size: 1.75em;
+  transition: opacity 350ms;
+}
+.honeycomb-cell_title > small {
+  font-weight: 300;
+  margin-top: 0.25em;
+}
+.honeycomb-cell::before,
+.honeycomb-cell::after {
+  content: '';
+}
+.honeycomb-cell::before,
+.honeycomb-cell::after,
+.honeycomb-cell_img {
+  top: -50%;
+  left: 0;
+  width: 100%;
+  height: 200%;
+  display: block;
+  position: absolute;
+  -webkit-clip-path: polygon(50% 0%, 100% 25%, 100% 75%, 50% 100%, 0% 75%, 0% 25%);
+  clip-path: polygon(50% 0%, 100% 25%, 100% 75%, 50% 100%, 0% 75%, 0% 25%);
+  z-index: -1;
+}
+.honeycomb-cell::before {
+  background: orange;
+  transform: scale(1.055);
+}
+.honeycomb-cell::after {
+  /* background: orange; */
+  opacity: 0.7;
+  transition: opacity 350ms;
+  -webkit-transition: opacity 350ms;
+}
+.honeycomb-cell:hover .honeycomb-cell_title {
+  opacity: 0;
+}
+.honeycomb-cell:hover .honeycomb-cell_img {
+  filter: grayscale(0%);
+}
+.honeycomb-cell:hover::before {
+  background: white;
+}
+.honeycomb-cell:hover::after {
+  opacity: 0;
+}
+.honeycomb_Hidden {
+  display: none;
+  opacity: 0;
+  width: 250px;
+  margin: 0 12.5px;
+}
+
+/****** Responsive *******/
+
+@media (max-width: 550px) {
+  .honeycomb-cell {
+    margin: 81.25px 25px;
+  }
+}
+
+@media (min-width: 550px) and (max-width: 825px) {
+  .honeycomb-cell:nth-child(3n) {
+    margin-right: calc(50% - 125px);
+    margin-left: calc(50% - 125px);
+  }
+
+  .honeycomb_Hidden:nth-child(3n + 5) {
+    display: block;
+  }
+}
+
+@media (min-width: 825px) and (max-width: 1100px) {
+  .honeycomb-cell:nth-child(5n + 4) {
+    margin-left: calc(50% - 275px);
+  }
+
+  .honeycomb-cell:nth-child(5n + 5) {
+    margin-right: calc(50% - 275px);
+  }
+
+  .honeycomb_Hidden:nth-child(5n),
+  .honeycomb_Hidden:nth-child(5n + 3) {
+    display: block;
+  }
+}
+@media (min-width: 1100px) {
+  .honeycomb-cell:nth-child(7n + 5) {
+    margin-left: calc(50% - 400px);
+  }
+
+  .honeycomb-cell:nth-child(7n + 7),
+  .honeycomb-cell:nth-child(7n + 5):nth-last-child(2) {
+    margin-right: calc(50% - 400px);
+  }
+
+  .honeycomb_Hidden:nth-child(7n + 7),
+  .honeycomb_Hidden:nth-child(7n + 9),
+  .honeycomb_Hidden:nth-child(7n + 11) {
+    display: block;
+  }
+}
+
+/* Members Select End*/
+
+/* LIGHTBOX START */
+a {
+  color: inherit;
+}
+
+a:hover {
+  color: #bbb;
+}
+
+.italic {
+  font-style: italic;
+}
+.small {
+  font-size: 0.8em;
+}
+
+/** LIGHTBOX MARKUP **/
+
+.lightbox {
+  /* Default to hidden */
+  display: none;
+
+  /* Overlay entire screen */
+  position: fixed;
+  z-index: 1;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+
+  /* A bit of padding around image */
+  padding: 1em;
+
+  /* Translucent background */
+  background: rgba(0, 0, 0, 0.7);
+  width: 100%;
+  overflow-x: scroll;
+  flex-wrap: wrap;
+}
+
+/* Unhide the lightbox when it's the target */
+.lightbox:target {
+  display: block;
+}
+
+.lightbox span {
+  /* Full width and height */
+  display: block;
+  width: 100%;
+
+  /* Size and position background image */
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: contain;
+}
+
+.lightbox .content {
+  font-size: 1.5rem;
+  text-align: left;
+}
+.lightbox .content img {
+  background-image: linear-gradient(to bottom right, gold, orange);
+  float: left;
+  margin-right: 6rem;
+  margin-bottom: 1rem;
+}
+
+table {
+  table-layout: fixed;
+  width: 100%;
+}
+
+td h1 {
+  word-wrap: break-word;
+  text-align: left;
+}
+td,
+th {
+  padding-left: 0rem;
+  padding-right: 0rem;
+  vertical-align: top;
+}
+
+.closeButton {
+  float: right;
+}
+
+@media (min-width: 300px) and (max-width: 999px) {
+  .lightbox .content {
+    font-size: 2vh;
+  }
+  .lightbox .content img {
+    width: 100%;
+    float: left;
+    margin-right: 1rem;
+  }
+  .lightbox h1 {
+    font-size: 3vh;
+  }
+}
+@media (min-width: 1000px) and (max-width: 1999px) {
+  .lightbox .content {
+    font-size: 2vh;
+  }
+
+  .lightbox .content img {
+    width: 40%;
+    float: left;
+    margin-right: 3rem;
+  }
+
+  .lightbox h1 {
+    font-size: 3vh;
+  }
+}
+
+@media (min-width: 2000px) {
+  .lightbox .content {
+    font-size: 2vh;
+  }
+
+  .lightbox .content img {
+    background-image: linear-gradient(to bottom right, gold, orange);
+    width: 40%;
+    float: left;
+  }
+
+  .lightbox h1 {
+    font-size: 3vh;
+  }
+}
+
+/* LIGHTBOX END */
